최적화 및 예외처리중
