+++ Syadeu/Database/ObArray.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0205bf127013679409717d0ad18aa00e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
