+++ Syadeu/Database/ObDictionary.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee1b3ff5c5e9f864bb044b9faed39e05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
