+++ Syadeu/Database/ObList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a788ab98d5b90e648a6ce7e950176cbc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
