+++ Syadeu/Database/ObservableBag.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9254de8b5f3b331428a6593457b4235b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
