+++ Syadeu/ECS/ECSPathBuffer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55243a538289f6c44b9e2388615ac5ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
