+++ Syadeu/ECS/ECSPathFinder.cs
@@ -1,0 +1,17 @@
+﻿using Unity.Entities;
+
+namespace Syadeu.ECS
+{
+    public struct ECSPathFinder : IComponentData
+    {
+        public int id;
+        public int agentTypeId;
+
+        public float maxTravelDistance;
+        public float overrideArrivalDistanceOffset;
+
+        public float radius;
+
+        public float speed;
+    }
+}
