+++ Syadeu/ECS/ECSPathFinder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61b63f154b9d36f47a0ff375d7624f1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
