+++ Syadeu/ECS/ECSPathObstacle.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Entities;
+
+namespace Syadeu.ECS
+{
+    public struct ECSPathObstacle : IComponentData
+    {
+        public int id;
+        public PathObstacleType type;
+    }
+}
