+++ Syadeu/ECS/ECSPathObstacle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c138d7fd6ea2a2e40b53c7b01450d93b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
