+++ Syadeu/ECS/ECSPathQuery.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8edb0c1589956244ad99b651794fbbc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
