+++ Syadeu/ECS/ECSPathSystemGroup.cs
@@ -1,0 +1,7 @@
+﻿using Unity.Entities;
+
+namespace Syadeu.ECS
+{
+    [UpdateAfter(typeof(ECSCopyTransformFromMonoSystem))]
+    public class ECSPathSystemGroup : ComponentSystemGroup { }
+}
