+++ Syadeu/ECS/ECSPathSystemGroup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60fbf86e73914aa42882d7606924fa35
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
