+++ Syadeu/ECS/Enum.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 105c4b5b5acc63c478442aa7287def93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
