+++ Syadeu/ECS/Enum/PathObstacleType.cs
@@ -1,0 +1,10 @@
+﻿namespace Syadeu.ECS
+{
+    public enum PathObstacleType
+    {
+        None, 
+        
+        Mesh,
+        Terrain,
+    }
+}
