+++ Syadeu/ECS/Enum/PathObstacleType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0770f186f58a5e248b1d1384a483d7fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
