+++ Syadeu/ECS/Enum/PathStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 474fb5d8505a7dc488dc20a87f8ac7fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
