+++ Syadeu/ECS/Enum/StraightPathFlags.cs
