+++ Syadeu/ECS/Enum/StraightPathFlags.cs.meta
