+++ Syadeu/ECS/Mono/ECSPathObstacleComponent.cs
