+++ Syadeu/ECS/Mono/ECSPathObstacleComponent.cs.meta
