+++ Syadeu/ECS/TransformSystem/ECSTransformFromMono.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46a599e2b7b96f3499e3ad4c11030bf3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
