+++ Syadeu/ECS/Utils.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6004403579b6277489e017b6c22369cd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
