+++ Syadeu/ECS/Utils/GeometryUtils.cs
