+++ Syadeu/ECS/Utils/PathUtils.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 81dd9f8107d97ed4392993dbd0a9154c
+guid: e81906da5596473498afce490c527a2f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
