+++ Syadeu/Mono/RenderManager.cs
@@ -20,7 +20,7 @@ namespace Syadeu.Mono
 
         public List<ManagedObject> ManagedObjects = new List<ManagedObject>();
 
-        private ObClass<Camera> MainCamera = new ObClass<Camera>(ObValueDetection.Changed);
+        private readonly ObClass<Camera> m_MainCamera = new ObClass<Camera>(ObValueDetection.Changed);
 
         internal Matrix4x4 CamMatrix4x4 { get; private set; }
 
@@ -36,10 +36,10 @@ namespace Syadeu.Mono
 
         private void Awake()
         {
-            MainCamera.Value = Camera.main;
-            MainCamera.OnValueChange += MainCamera_OnValueChange;
+            m_MainCamera.Value = Camera.main;
+            m_MainCamera.OnValueChange += MainCamera_OnValueChange;
 
-            StartUnityUpdate(UnityUpdate());
+            //StartUnityUpdate(UnityUpdate());
         }
 
         private void MainCamera_OnValueChange(Camera current, Camera target)
@@ -47,40 +47,48 @@ namespace Syadeu.Mono
             CamMatrix4x4 = GetCameraMatrix4X4(target);
         }
 
-        private IEnumerator UnityUpdate()
-        {
-            while (m_Instance != null)
-            {
-                if (MainCamera == null) MainCamera.Value = Camera.main;
-                CamMatrix4x4 = GetCameraMatrix4X4(MainCamera.Value);
+        //private IEnumerator UnityUpdate()
+        //{
+        //    while (m_Instance != null)
+        //    {
+        //        if (MainCamera == null) MainCamera.Value = Camera.main;
+        //        CamMatrix4x4 = GetCameraMatrix4X4(MainCamera.Value);
 
-                if (m_WaitForManaged.Count > 0)
-                {
-                    int count = m_WaitForManaged.Count;
-                    for (int i = 0; i < count; i++)
-                    {
-                        if (m_WaitForManaged.TryDequeue(out var controller))
-                        {
-                            ManagedObjects.Add(new ManagedObject(controller));
-                        }
-                    }
-                }
+        //        if (m_WaitForManaged.Count > 0)
+        //        {
+        //            int count = m_WaitForManaged.Count;
+        //            for (int i = 0; i < count; i++)
+        //            {
+        //                if (m_WaitForManaged.TryDequeue(out var controller))
+        //                {
+        //                    ManagedObjects.Add(new ManagedObject(controller));
+        //                }
+        //            }
+        //        }
 
-                for (int i = 0; i < ManagedObjects.Count; i++)
-                {
-                    if (ManagedObjects[i].Controller.IsInvisible)
-                    {
-                        ManagedObjects.RemoveAt(i);
-                        i--;
-                        continue;
-                    }
+        //        for (int i = 0; i < ManagedObjects.Count; i++)
+        //        {
+        //            if (ManagedObjects[i].Controller == null)
+        //            {
+        //                throw new CoreSystemException(CoreSystemExceptionFlag.Render,
+        //                    "RenderController를 매니저에서 컨트롤하는 도중 파괴되었습니다. " +
+        //                    "파괴되기전 먼저 매니저 관리를 해제하세요");
+        //            }
+        //            if (ManagedObjects[i].Controller.IsInvisible)
+        //            {
+        //                ManagedObjects.RemoveAt(i);
+        //                i--;
+        //                continue;
+        //            }
 
-                    if (i != 0 && i % 500 == 0) yield return null;
-                }
+        //            // 추가 작업
 
-                yield return null;
-            }
-        }
+        //            if (i != 0 && i % 500 == 0) yield return null;
+        //        }
+
+        //        yield return null;
+        //    }
+        //}
         private void OnDestroy()
         {
             for (int i = 0; i < ManagedObjects.Count; i++)
@@ -91,11 +99,11 @@ namespace Syadeu.Mono
             ManagedObjects.Clear();
         }
 
-        private readonly ConcurrentQueue<RenderController> m_WaitForManaged = new ConcurrentQueue<RenderController>();
-        internal void AddRenderControl(RenderController controller)
-        {
-            m_WaitForManaged.Enqueue(controller);
-        }
+        //private readonly ConcurrentQueue<RenderController> m_WaitForManaged = new ConcurrentQueue<RenderController>();
+        //internal void AddRenderControl(RenderController controller)
+        //{
+        //    m_WaitForManaged.Enqueue(controller);
+        //}
 
         /// <summary>
         /// 렌더링 규칙을 적용할 카메라를 설정합니다.
@@ -104,7 +112,7 @@ namespace Syadeu.Mono
         /// <param name="offset"></param>
         public static void SetCamera(Camera cam)
         {
-            Instance.MainCamera.Value = cam;
+            Instance.m_MainCamera.Value = cam;
         }
 
         //// 이거 젤터 전용
